Rework MailList.available_languages so that we don't need to use a PickleType column in the database for this list of strings.  We use SQLAlchemy's many-to-many relationship, however because of this, you cannot simply append new unicodes to .available_languages.  You need to wrap the language code in a Language instance and append that instance to the list.
In order to handle this, I added a property MailList.language_codes which
returns a list of the code strings (not Language instances).  Also new are
MailList.set_languages() for setting (i.e. overriding) the set of available
languages for the list; and add_language() which takes a single language code,
wraps it, and appends it.  The code does not and should not use
.available_languages directory any more.

MailList.GetAvailableLanguages() is removed.  The 'available_languages' column
is removed from the Listdata table.

Add a getValue() to Mailman.Gui.Language in order to unwrap the language codes
stored in the database's association table.  Modify _setValue() to do the
wrapping.

In dbcontext.py, don't import * from the sqlalchemy package.  It contains a
'logging' name which is not the standard Python logging package.  I also added
essentially a bag of attributes class called Tables which will hold references
to all the SA tables that are created.  Update the make_table() API to take an
instance of Tables.

Added a close() method to DBContext.  This is needed for the updated unit test
suite.

Changed bin/import.py so that when available_languages is being set, it calls
MailList.set_languages() instead of trying to set that attribute directly.

Updated some language idioms while I was at it.

More eradication of mm_cfg in favor of the config object and the Defaults
module.

In testall.py, call initialize() instead of loginit.initialize().

Promote MAX_RESTARTS into a Defaults.py.in variable.  This is because the unit
tests will knock that value down to something not so annoying should one of
the qrunner-required tests traceback.

Several other important changes to the unit test suite (which now completely
succeeds again!):

- Set the uid and gid of the temporary mailman.cfg and tmp*.db files to the
  Mailman user and group as specified in the config object.

- Make sure that all of the tests point to a SQLite database file that was
  created with the tempfile module.  This way we don't pollute our main
  database with data that is getting created during the unit tests.

- In the TestBase.setUp() method, be sure to close the existing dbcontext,
  clear out the mappers, and then reconnect the dbcontext with the new
  SQLALCHEMY_ENGINE_URL pointing to the tempfile.  However, we don't need to
  reload the MailList instance any more.

- Make all tests work, except for the tests that require crypt.  That upgrade
  path will not be available in this version of Mailman.