api_lock(): When locking the MailList object, tell the SQLAlchemy session to expire the object.  This way, when the MailList attributes are next accessed, the ORM will reload them from the database, getting any new values possibly set in other processes.
This works better than trying to use always_refresh=True on the mapper, or
trying to do a reload() because both of those approaches blow away locks.  I'm
not sure why this, but I suspect that it's because the identity map is handing
us back a different object, rather than invalidating the object's attributes.