Moved the session.expire() to the MailList.Load() method, via DBContext.api_load().  This is where it really ought to be based on the internal semantics of .Load()/.Lock().
MailListMapperExtension.populate_instance(): Checking the state of the isnew
flag is not sufficient to know whether the MailList object we're getting is
brand spankin' new or not.  It turns out that when we session.expire() the
MailList object, the next time SA loads this from the db, the
populate_instance() will get called with isnew=True, even though the object
really isn't new.  Instead, check to make sure InitTempVars() isn't
incorrectly called twice.  Note that I might move this test, but I wanted to
check in something that works, and then see if this is what we expect from the
SA guys (this flag appears underdocumented).

LockFile.py: Add some additional debugging.