Update the IUser interface and tests, specifically as it relates to preferences.  IAddresses, IUsers, and IMembers all get preferences by default, althoughthe attributes of these preferences are None by default.
IMailingLists don't get preferences by default though; because these live in
the user database, we can't cross-polinate them in the mailing lists.  We'll
figure something out later for these.

IUser.register(): Add this method which registers and links an address to the
user.

Allow EnumType database columns to accept and return Nones.  This is useful
for when the columns are not defined NOT NULL.

Update doctests.

Removed teh hide_address preference.  I can't think of a reason not to want to
hide addresses for everyone.