Convert the tests for the CalcRecips handler to doc tests.  There are some XXX's in the doc test because digest recipients aren't tested (though those may go in a different doctest), and neither are urgent messages.  This latter is for the same reason that the Approved handler is not yet tested; which password do you use in that header?
The CalcRecips tests would also seem the natural place to test the
receive_list_copy preference, but that actually gets processed in the
AvoidDuplicates handler, so it isn't tested here.

Add delivery_status (of type enum DeliveryStatus) to preferences.  I'm
not entirely sure that's the right place for it, but it lets me finish
converting the test for now.

Expose the rest of the preferences through the IMember interface.