Convert the Switchboard test in test_runners.py to a doctest.  Add an ISwitchboard interface and modernize the Python code in the Switchboard.py implementation.
The SAVE_MSGS_AS_PICKLES option is removed.  Messages are always saved as
pickles unless the metadata '_plaintext' key is present, though this should
eventually go away too.

In testall.py, put the entire VAR_PREFIX in a temporary directory.  This helps
the switchboard tests by not mixing their data with the installation's queue
directories.  The Configuration object now also ensures that all the queue and
log directories exist -- one more step on the road to getting rid of the
autoconf mess.