Convert the rest of test_runners.py to doctests; even though incomplete, they test everything the old unit tests tested.  There are XXX's left in the doctests as reminders to flesh them out.
Change the NNTP_REWRITE_DUPLICATE_HEADERS to use proper capitalization.

Revert a change I made in the conversion of the Switchboard class:
Switchboard.files is no longer a generator.  The Runner implementation is
cleaner if this returns a concrete list, so that's what it does now.  Update
the tests to reflect that.

The Runner simplifies now too because it no longer needs _open_files() or the
_listcache WeakValueDictionary.  The standard list manager handles all this
now, so just use it directly.

Also change the way the Runner sets the language context in _onefile().  It
still tries to set it to the preferred language of the sender, if the sender
is a member of the list.  Otherwise it sets it to the list's preferred
language, not the system's preferred language.  Removed a conditional that
can't possibly happen.