Support for case-preserving addresses.  When an Address is given an email address that is not lower cased, the original, case-preserved version is store on the '_original' attribute.  The lower-cased version is always used as the key and thus always stored on the 'address' attribute.  The IAddress interface is given a new 'original_address' property which returns the case-preserved version.
Address's __str__() and __repr__() are similarly modified.  The former always
includes the case-preserved address; the latter does too, but now also
includes the lower-cased 'key' email address (along with the object's id).

Searching for an address always does so on the lower-cased version.

Test suite is updated as necessary.  Also, I'm adding the
REPORT_ONLY_FIRST_FAILURE doctest flag so that it's easier to debug doctest
failures without having pages of problems to scroll through.