Convert failing test_message.py to doctests bounces.txt and message.txt, which of course now succeed.
Rename Bouncer.py's BounceMessage() method to bounce_message() and remove the
'msgdata' parameter, which wasn't being used.  Change the RejectNotice
exception class to expose .notice directly, as there's no reason for this to
be an accessor or property.

Move the coverage.py installation to the install-packages target instead of
the install-other target, so that it only gets installed once the pythonlib
directory is created.