OMGW00T: After over a decade, the MailList mixin class is gone!  Well, mostly.  It's no longer needed by anything in the test suite, and therefore the list manager returns database MailingList objects directly.  The wrapper cruft has been removed.
To accomplish this, a couple of hacks were added to the Mailman.app
package, which will get cleaned up over time.  The MailList module
itself (and its few remaining mixins) aren't yet removed from the tree
because some of the code is still not tested, and I want to leave this
code around until I've finished converting it.