Reorganize the qrunner infrastructure.  First, the package has been renamed from Mailman.Queue to Mailman.queue (note the case change to be more PEP 8 compliant).
The Switchboard and Runner classes have been moved into the package
__init__.py and the previous class modules have been removed.

The switchboard cache is removed; I don't think it was ultimately buying us
much.  Now, just import the Switchboard class and instantiate it directly.

Added an IRunner interface.

Renamed the ArchRunner to ArchiveRunner.

bin/qrunner and bin/mailmanctl are updated accordingly.  For the former, it no
long accepts -r=All to run all qrunners.  You can still use the short name
(e.g. --runner=incoming) to run the built-in queue runners, but this design
will eventually allow for plugin qrunners by allowing them to be run
specifying the full package path to the class.  It also now accepts a leading
dot to indicate a qrunner class relative to the Mailman.queue package.