Fix two doctests: ack-headers and acknowledgment.
This hacks around an apparent bug in the email package where if you parse a
unicode message string, you still end up getting 8-bit strings out of the
headers, and probably payloads.

The hack is to override Mailman.Message.Message.__getitem__() to force the
header value returned to a Unicode.  It must be ASCII but this is required
anyway by RFC 2822.  It's not perfect, but it lets us get farther without
forcing a detour into fixing the email package.

Other changes:

- Fix the Address table's references, and also update the subscribe() query.
- Fix the Member table's references and add a __init__().
- Fix Roster's get_member() query.
- Fix the Enum class's variable_class attribute.
- UserManager.create_user() has to use Unicodes for real_name.