More changes to rules and chains.
Now a link has a rule, action, chain, and function, not all of which needs to
be specified.  The action is a LinkAction enum adn specifies what to do should
the rule match.  The use of the chain or function depends on what the action
is.

Several interface changes now make it easier to jump to other chains, push
(i.e. detour) to chains, etc.  Rules can also now specify that they should not
be recorded in X-* headers.

Added a TruthRule which always matches.