Reorganization.
Create a Mailman.chains package and move a bunch of the Mailman.apps.chains
classes to separate modules under this package.

Also, separate out iteration from chain management.  In other words, IChain
and IChainIterator are separate, and the latter is retrieved from the former
by the .get_links() method.  This latter takes the mailing list, message, and
metadata dictionary so that links can be targetted specifically to the task at
hand.