SpamDetect is gone, so the chains/rules implementation experiment is deemed a success and will now be merged into the trunk.
Move the Truth rule into the built-in rules package in a separate module, and
add a test.

Modify IChainLink so that the rule and chain attributes are not names but
indeed the actual ILink or IChain object directly.  Update the
chains.process() function accordingly.  Remove the IChain.get_rule() method.

Don't derive BuiltInChain from Chain and don't make it an IMutableChain.  It's
now just an IChain, and is implemented concretely.

Refactor the HeaderMatchChain and friends so that it can be used with both the
global HEADER_MATCHES variable and the list-specific header_matches variable,
which has exactly the same semantics.  Oh yeah, get rid of the list's
header_filter_rules attribute and replace it with header_matches so that the
semantics match, it's easy to explain, and it's all nice and clean.