Convert IncomingRunner to use the new chains disposition architecture.  move the big explanatory text at the beginning of incoming.py to a doctest called OVERVIEW.tt (which doesn't actually contain any tests yet -- it's documentation though).
Added a doctest for the incoming runner, though this will be fleshed out in
more detail next.

Mailman.Post renamed to Mailman.inject, and simplified.  We don't need its
command line script behavior because that is now handled by bin/inject.

Add a 'start_chain' attribute to mailing lists.  This names the chain that
processing of messages for that list begins with.

We were inconsistent in the use of the 'no reply' address attribute.  It's now
always 'no_reply_address'.

Update the smtplistener helper with lessons learned about how to suppress
bogus asyncore error messages.  Also, switch to using a maildir mailbox
instead of an mbox mailbox.