Added a test for the built-in pipeline.  Fixed some broken handler names in the built-in pipeline.
Fixed DEFAULT_SUBJECT_PREFIX to take $-names instead of %-names, although I'm
not entirely sure the %%d stuff still works (we need a test for this).

Added IMailingList.real_name attribute and added this to the default style.  A
column for this was in the database but not in the storm object.  Also
re-enabled the style's subject_prefix attribute.

Moved some of the digest test functions into Mailman.tests.helpers.