Rework the basic infrastructure for qrunner process control.  Split out the functionality of mailmanctl into a separate master watcher script.  mailmanctl has not yet been updated but that'll happen next.
Fix DELIVERY_MODULE to name a handler instead of a module.

Change make_instance to use pkg_resources instead of module.__file__.

Change the qrunner and master processes coordination so that the qrunners are
not restarted on SIGINT, because otherwise C-c just doesn't work.  Now SIGUSR1
is how we'll implement 'mailman restart'.

Add a database commit so that initializing the schema doesn't lock the sqlite
database.  Also, don't try to initialize the schema if the tables already
exist.  Use some sqlite magic to do this test.

Move mailman.cfg.in into a new package Mailman/extras inside the tree.  Also,
MAILMAN_UID and MAILMAN_GID should be integers not strings.

Convert the command runner to use an IHandler instance instead of handler
module.  Similarly for the outgoing runner, DELIVERY_MODULE now names an
IHandler instance instead of a handler module.