Add a working (though not yet complete) test for the LMTP runner.
Refactor the TestableMaster class so that doctests don't need to do all the
threading and event fiddling themselves.  Hide all that in the test class.

In bin/master, add the -r/--runner option to allow overriding the set of queue
runners to start from the command line.  This is much more convenient than
fiddling the config file (which is still supported), and it allows us to start
a subset of the runners from the TestableMaster class.

Refactor the calculation of queue runner shortcut names into
Configuration.add_qrunner().  This way, it's easy to share between bin/qrunner
and bin/master.

Use pkg_resource in bin/testall to create the test configuration file.  We're
still using mailman.__file__ as the target to os.walk() though, so this
conversion isn't complete.

In bin/testall, update to the new convention of putting .options and
.arguments on the parser instance, so we only need to pass back one thing.

In test_documentation.py, use config.verbosity instead of
config.options.verbosity.  Wind that through to bin/testall while we're at
it.

Update loginit.py so that defaults for all the logging options can be found in
a [*] section.  This is better than [DEFAULT] because the latter requires
%-interpolation and still requires each sublogger to be specified explicitly.
Now we just set values in the [*] section and have them apply to all loggers.

Allow for passing bin/testall's -v and -e options to any subprocesses that get
created, e.g. via the TestableMaster.  Now testall will write a logging.cfg
file with a [*] entry as appropriate.

It's DEFAULT_DATABASE_URL now, not SQLALCHEMY_ENGINE_URL.

StockDatabase._reset() requires that the store be rolled back before it's
reset.  Otherwise if there are outstanding transactions, the reset will fail
with an OperationalError.

Update the LMTPRunner.  Have it return a 501 error if the message has defects.
Most defective messages are spam.  Make the LMTPRunner startable and
stoppable.  Just because it's asyncore based doesn't mean that's not possible
<wink>.

Enable the LMTP runner in tests.