Get rid of one_last_digest.  Move this into a separate OneLastDigest table. Make it explicit to get the last digest (I could imagine an user selecting not to).  Actually add tests for this.
Start to get rid of 'from storm.locals import *' in favor of more specific
imports.  Start to use Store.of() instead of config.db.store where we can.

Rework (delivery)MemberRosters to use the Member attributes.  This ensures
that the layered lookup happens at the expense of query optimization.

Sundry and various cleanups.