Clean a few more lints.
Add get_mailing_lists() which is used just for the web interface.  Because of
a bug in lazr.restful, this cannot be a generator.

Similar change in IDomainSet.

Instrument IListManager to be vended through the api.

The REST server must be run in a separate process since SQLite does not like
objects created in one thread to be used in another thread.  Note that this
breaks the domain.txt test, but domains really need to be in the database
anyway.