Oh FFS.  lazr.restful 0.9.18 totally broke our shit by introducing multiversioned web services.  In concept, that's a great idea, but in (current as of 0.9.18) practice it sucks because it /forces/ us to adopt multiversions when really I could care less.  After gobs of painful experimentation, this passes all the tests.  I guess that means it's right <wink>.
Another, more reasonable fix is to re-order when logging is initialized.  This
is moved to the start of initialize_2() because in the test environment, we
don't want to initialize the loggers until after the test configuration has
been pushed.  Otherwise, because of the changes to support the FHS, logging
will go to the wrong place.  This really only affects tests, since in
operational mode, initialize always happens immediately one after another.