More fixes to the start up procedure.  The root cause of the fix is that when var_dir in the config file is a relative path, the bin/master and bin/qrunner subprocesses end up creating another hierarchy under $VAR_DIR.  This is because 'bin/mailman start' chdirs to $VAR_DIR before starting master.
The fix involves putting the absolute path to $VAR_DIR in the $MAILMAN_VAR_DIR
environment variable.  If the configuration file reader sees this, it uses the
value instead of $var_dir from the configuration file.

Another change is to remove qrunner's --subproc/-s option and stick this in
the environment as well, using $MAILMAN_UNDER_MASTER_CONTROL.

Other fixes here include:
* Deciphering the master lock file contents correctly
* Reformatting the error message so that it fits in 79 columns when prefixed
  by argparse error info.