encrypt_password(): New convenience function for ensuring that a password is both encrypted according to a scheme, and a bytes object.
add_member(): Use encrypt_password().

cli_members: Give the user a default, user-friendly password.  Of course, this
will be encrypted so it can't be retrieved, but it can be reset.

Passwords are stored as bytes objects, not unicode now.

ConfigLayer: Set the default test password scheme to cleartext.

General test repair.